Set MaxIdleConnsPerHost for Node.searchClient
For as long as http.Transport does not have timeouts for cached
connections, MaxIdleConnsPerHost should be set quite high, since
ElasticSearch can obviously handle it.

At some point, this should probably be exposed as a configuration
option, but 250 seems like a sane default for sites with a high-ish
number of concurrent connections.